requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
Flask-Mail==0.9.1
tenacity
flask-pymongo==2.3.0
pymongo==4.3.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
langdetect==1.0.9
python-dateutil==2.8.2
cerberus==1.3.4
flask-swagger-ui==4.11.1
gunicorn==20.1.0
greenlet==2.0.2
pytest==7.4.0
pytest-cov==4.1.0
requests-mock==1.11.0
freezegun==1.2.2
flake8==6.0.0
black==23.7.0
pylint==2.17.4
debugpy==1.6.7
google-ai-generativelanguage==0.6.15
transformers==4.52.4
grpcio-status==1.73.0
protobuf==3.20.3
grpcio==1.73.0
